feat: convert to difficulty class system & reduce duplication in frontend (#14)

+++ backend/app/agents/trust_scoring_agent.py
@@ -16,7 +16,7 @@ MAX_MESSAGE_HISTORY = 10
 
 
 class TrustCalculatorAgentOutput(BaseModel):
-    score: float
+    score: int
     reason: str
